backend/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signalsentry-backend"
version = "0.1.0"
description = "FastAPI backend for SignalSentry"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "pydantic-settings>=2.2.0",
    "python-multipart>=0.0.9",
    "alembic>=1.13.1",
    "reportlab>=4.0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["app"]
